+++ boca/__init__.py
@@ -1,2 +1,3 @@
 from .cli import create_cli
 from .version import __version__
+from .utils import call_command